cdo: move subsystem definition generation to 'xlnx_pm' area
 - Introduce a new subdir "xlnx_pm" in assists/* which contains xilinx
 specific files for handling the generation of subsystem/protection cdo
 configuration
